修改了数据库

+++ blog/blog-core/src/main/resources/sql/vote.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS `vote`;
 CREATE TABLE `vote` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
-  `user_id` INT(11),
-  `blog_id` INT(11),
-  `status` INT(1) DEFAULT 1,
+  `id` INT NOT NULL AUTO_INCREMENT,
+  `user_id` INT,
+  `blog_id` INT,
+  `status` TINYINT DEFAULT 1,
   `create_time` TIMESTAMP,
   PRIMARY KEY (`id`),
   FOREIGN KEY (`user_id`) REFERENCES `user` (`id`),
